requirements.txt
numpy
pandas
scikit-learn
rdkit-pypi
torch
gpytorch
pytest
tqdm
selfies